requirements.txt
# PS5-Bilibili-Danmaku Python依赖包
# 使用: pip install -r requirements.txt

# Web框架
flask>=2.3.0

# HTTP客户端
requests>=2.31.0
aiohttp>=3.9.0

# 二维码生成（可选，用于扫码登录）
qrcode>=7.4.0
Pillow>=10.0.0
